backend/app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
psycopg[binary]
alembic
python-dotenv
